refactor: rename timeline to feed
