+++ src/screens/index.js
@@ -1,8 +1,8 @@
 import OnBoarding from './onboarding/OnBoarding';
 import Splash from './splash/Splash';
-import Timeline from './timeline/Timeline';
+import Feed from './feed/Feed';
 import CreatePost from './createPost/CreatePost';
 import Favorits from './favorits/Favorits';
 import MyPosts from './myPosts/MyPosts';
 
-export {OnBoarding, Splash, Timeline, CreatePost, Favorits, MyPosts};
+export {OnBoarding, Splash, Feed, CreatePost, Favorits, MyPosts};
